% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statfi_data.R
\name{get_statfi_data.default}
\alias{get_statfi_data.default}
\title{Import pxweb data as data.frame}
\usage{
get_statfi_data.default(url, query)
}
\arguments{
\item{query}{}
}
\description{
Import pxweb data as data.frame
}
