% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfi_urls.R
\name{open_statfi_qui}
\alias{open_statfi_qui}
\title{Open table in QUI PxWeb}
\usage{
open_statfi_qui(x = "", db_list_name = NULL)
}
\arguments{
\item{x}{character, url or table code. If table code, set db_list_name}

\item{db_list_name}{character, database of the table}
}
\description{
Open table in QUI PxWeb
}
\examples{

  open_statfi_qui("StatFin__vrm__muutl/statfin_muutl_pxt_119z.px/")

}
