% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfi_urls.R
\name{table_code_to_url}
\alias{table_code_to_url}
\title{Get url from table code}
\usage{
table_code_to_url(x, db_list_name, with_base_url = FALSE)
}
\arguments{
\item{x}{A url}

\item{db_list_name}{chr}

\item{with_base_url}{logical. Defaults to \code{FALSE}}
}
\description{
Inverse of get_table_code. However, a table code is not sufficient information
for url so this has to use a db_list to map table codes to urls.
}
\details{
TODO: if db_list_name not given, url should be retrievable from statfi.
}
